+++ packages/web/src/components/ExpansionItem.vue
